Support for multiple data- or sequence- sections as the message body.
For sending a message use the message.data_sections() (note plural)
message.sequence_sections() alternatives to data_section()/sequence_section()
and pass in an array of values for these sections.

When a message is received with multiple data- or sequence- sections for the
body, the body will be an object with 'multiple' set to true, 'typecode' set
to the numeric typecode of the section and 'content' set to an array containing
the contents of the sections.

This fixes isseu #45