Ensure connection, sessions and links return false for is_open() when disconnected.
This changes splits the reset() method into two stages: _disconnect()
and _reconnect(). The former is always invoked. The latter is invoked
only when reconnecting. The desired state of endpoints (i.e. whether
or not they were open at the point of disconnection) is tracked
internally so as to be able to restore sessions and links as required
on reconnect.