Update README.md and add tests to prove it correct
I had a crash with session_error due to close of RabbitMQ broker and
did not find documentation. Also, when to create links and call send
was unclear.

I remove open_receiver and open_sender out of connection_open, because it
is called again after reconnecting. Better not to create new links on every
reconnect.

Session object and events are now added and missing events of other objects
as well. The send method now mentions the limit of unsent deliveries on
session.